refactor more paths
